requirements.txt
aiohttp
apipkg
appnope
asn1crypto
async-timeout
atomicwrites
attrs
autopep8
backcall
beautifulsoup4
bleach
boto
CacheControl
cachy
certifi
cffi
chardet
cleo
Click
clikit
colorama
coverage
cryptography
cycler
dash
dash-core-components
dash-html-components
dash-renderer
dash-table
decorator
defusedxml
entrypoints
execnet
fire
flake8
Flask
Flask-Compress
future
gevent
greenlet
grequests
haasomeapi
html5lib
idna
importlib-metadata
ipdb
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
Jinja2
jsonpickle
jsonschema
jupyter-client
jupyter-core
keyring
kiwisolver
lockfile
MarkupSafe
matplotlib
mccabe
mistune
more-itertools
msgpack
multidict
nbconvert
nbformat
nose
notebook
numpy
packaging
pandas
pandocfilters
parameterized
parso
pastel
pdfminer.six
pexpect
pickleshare
pkginfo
plotly
pluggy
poetry
portalocker
prometheus-client
prompt-toolkit
ptyprocess
py
pycodestyle
pycparser
pycryptodome
pyflakes
Pygments
pylev
PyMySQL
pyOpenSSL
pyotp
pyparsing
pyrsistent
pytest
pytest-cov
pytest-forked
pytest-html
pytest-metadata
pytest-ordering
pytest-rerunfailures
pytest-xdist
python-dateutil
python-fire
pytz
pyzmq
requests
requests-toolbelt
retrying
scipy
selenium
seleniumbase
Send2Trash
setuptools-scm
shellingham
six
sortedcontainers
soupsieve
spicy
SQLAlchemy
termcolor
terminado
testpath
tomlkit
tornado
tqdm
traitlets
urllib3
wcwidth
webencodings
Werkzeug
widgetsnbextension
yarl
zipp
limits
